refactor: update account handling to use event-driven architecture with strategies

+++ src/main/java/br/com/giovanecarvalho/accountchallenge/dto/AccountDto.java
@@ -1,4 +1,6 @@
 package br.com.giovanecarvalho.accountchallenge.dto;
 
-public record AccountDto(String id, int balance) {
+import java.math.BigDecimal;
+
+public record AccountDto(String id, BigDecimal balance) {
 }
