+++ src/main/java/br/com/giovanecarvalho/accountchallenge/dto/EventRequest.java
@@ -1,4 +1,6 @@
 package br.com.giovanecarvalho.accountchallenge.dto;
 
-public record EventRequest(String type, String origin, String destination, int amount) {
+import java.math.BigDecimal;
+
+public record EventRequest(String type, String origin, String destination, BigDecimal amount) {
 }
